+++ core/src/main/java/org/apache/calcite/avatica/remote/LocalService.java
@@ -90,6 +90,8 @@ public class LocalService implements Service {
     }
 
     Meta.Signature signature = resultSet.signature;
+    // TODO Revise modification of CursorFactory see:
+    // https://issues.apache.org/jira/browse/CALCITE-4567
     Meta.CursorFactory cursorFactory = resultSet.signature.cursorFactory;
     Meta.Frame frame = null;
     int updateCount = -1;
@@ -97,20 +99,24 @@ public class LocalService implements Service {
 
     if (resultSet.firstFrame != null) {
       list = list(resultSet.firstFrame.rows);
-      switch (cursorFactory.style) {
-      case ARRAY:
+      if (list.isEmpty()) {
         cursorFactory = Meta.CursorFactory.LIST;
-        break;
-      case MAP:
-      case LIST:
-        break;
-      case RECORD:
-        cursorFactory = Meta.CursorFactory.LIST;
-        break;
-      default:
-        cursorFactory = Meta.CursorFactory.map(cursorFactory.fieldNames);
+      } else {
+        switch (cursorFactory.style) {
+        case ARRAY:
+          cursorFactory = Meta.CursorFactory.LIST;
+          break;
+        case MAP:
+        case LIST:
+          break;
+        case RECORD:
+          cursorFactory = Meta.CursorFactory.map(cursorFactory.fieldNames);
+          break;
+        default:
+          throw new IllegalStateException("Unknown cursor factory style: "
+              + cursorFactory.style);
+        }
       }
-
       final boolean done = resultSet.firstFrame.done;
 
       frame = new Meta.Frame(0, done, list);